Refactored app.jsx to separate out reviewParent.jsx, finished drop down animation
